refactor: mailer: splitting code.
