+++ mailer/MailerList.js
@@ -8,7 +8,7 @@ const logger = rootRequire('logger').prefix('MAILER').prefix('LIST');
 
 const Mailer = require('./Mailer.js');
 
-const _ = require('lodash');
+const sync = require('./MailerList.sync.js');
 
 class MailerList {
   constructor () { }
@@ -388,242 +388,8 @@ class MailerList {
         );
       });
     }
-    const providerName = mailerProvider.getName();
-    logger.log(`[startSync] provider ${providerName}`);
-    //
-    // we need to compute a DIFF between
-    //  List <-> Subscribers               (MailerAssoListsSubscribers)
-    //  List <-> [Subscribers,Providers]   (MailerAssoListsSubscribersProviders)
-    //
-    // reporting the diff can be quite long, so we write in the database our "process number"
-    // & test if this process number still exists during the process.
-    //
-    let syncId = Math.round(Math.random() * 1000000);
 
-    return this.getAssoProvider(mailerProvider)
-      .then(assoListProvider => {
-        // check if sync process is not already in progress.
-        if (assoListProvider.pApiStatus && assoListProvider.pApiStatus.id) {
-          throw new Error('sync already started');
-        }
-        // saving new syncId
-        return assoListProvider.updatePApiStatus({ sync: { id: syncId, startedAt: new Date() }});
-      })
-      .then(assoListProvider => {
-        const l = logger;
-        // to sync, we need to :
-        // - fetch all asso ListSubscriber
-        // - compare these asso.state to all asso ListSubsciberProvider state
-        //
-        setTimeout(() => {
-          /*
-           *
-           *
-           *   W O R K E R       B E G I N
-           *
-           * (we are in the worker id=syncId)
-           *
-           */
-          const logger = l.prefix('startSync').prefix('worker').prefix(syncId);
-
-          logger.log(`START`);
-
-          // the worker is async.
-          // the worker check periodicaly if syncId is still his own in the database
-          //  if not => it stops the processing.
-          //
-          Q.all([
-            sqldb.MailerAssoListsSubscribers.findAll({
-              where:{listId:this.getId()}
-            }),
-            sqldb.MailerAssoListsSubscribersProviders.findAll({
-              where:{listId:this.getId(),providerId:mailerProvider.getId()}
-            })
-          ])
-          .then(([assoLS, assoLSP]) => {
-            // assoLS  = assoListsSubscribers
-            // assoLSP = assoListsSubscribersProviders
-            const assoLSSubscribersIds = assoLS.map(o => o.subscriberId);
-            const assoLSPSubscribersIds = assoLSP.map(o => o.subscriberId);
-
-            logger.log(`assoLS: ${assoLSSubscribersIds.join(',')}`);
-            logger.log(`assoLSP: ${assoLSPSubscribersIds.join(',')}`);
-
-            const toCreate = _.difference(assoLSSubscribersIds, assoLSPSubscribersIds)
-              .map(subscriberId => { return {
-                listId: this.getId(),
-                subscriberId: subscriberId,
-                providerId: mailerProvider.getId()
-              }; });
-
-            logger.log(`creating: ${toCreate.length} id`);
-
-            return sqldb.MailerAssoListsSubscribersProviders.bulkCreate(
-              toCreate, { individualHooks: true }
-            );
-          })
-          .then(() => {
-            logger.log(`fetching data`);
-
-            return Q.all([
-              sqldb.MailerAssoListsSubscribers.findAll({
-                where: {listId: this.getId(), state: 'ACTIVE', disabled: false},
-                include: [{model: sqldb.MailerSubscriber, required: true, as: 'subscriber'}]
-              }),
-              sqldb.MailerAssoListsSubscribersProviders.findAll({
-                where: {listId: this.getId(), providerId: mailerProvider.getId()},
-                include: [{model: sqldb.MailerSubscriber, required: true, as: 'subscriber'}]
-              })
-            ]);
-          })
-          .then(([assoLS, assoLSP]) => {
-            // assoLS  = assoListsSubscribers
-            // assoLSP = assoListsSubscribersProviders
-
-            //
-            // list of subscribers to activate in the providers
-            //  <=> assoListsSubscribersActive without corresponding assoListsSubscribersProviders
-            //        with state !== null.
-            //
-            // the algorithm is :
-            //  - we flag the asso "list suscriber provider" to P-ACTIVE (pending)
-            //  - we create the subscriber using pApi
-            //  - if ok,  we flag the user ACTIVE
-            //  - if nok, we flag the user E-ACTIVE (error)
-            //
-            const idToAssoLSP = { /* subscriberId : assoLSP */ };
-            const idToAssoLS = { /* subscriberId : assoLS */ };
-
-            assoLSP.forEach(asso => idToAssoLSP[asso.subscriberId] = asso);
-            assoLS.forEach(asso => idToAssoLS[asso.subscriberId] = asso);
-
-            logger.log(`data fetched`);
-
-            // toCreate & toDelete are assoLSP models.
-            const assoLSPtoCreate = assoLS.filter(asso =>
-              asso.subscriberId &&
-              idToAssoLSP[asso.subscriberId].subscriberCanBeCreatedInProviderAPI()
-            ).map(asso => idToAssoLSP[asso.subscriberId]);
-            const assoLSPtoDelete = assoLSP.filter(asso =>
-              typeof idToAssoLS[asso.subscriberId] === 'undefined' &&
-              asso.susbcriberCanBeDeletedInProviderAPI()
-            );
-            const pApi = mailerProvider.getAPIInterface();
-
-            logger.log(`assoLSPtoCreate = [${assoLSPtoCreate.map(a=>a.subscriberId+':'+a.subscriber.referenceEmail).join(',')}]`);
-            logger.log(`assoLSPtoDelete = [${assoLSPtoDelete.map(a=>a.subscriberId+':'+a.subscriber.referenceEmail).join(',')}]`);
-
-            return this.getAssoProvider(mailerProvider)
-              .then(assoListProvider => {
-                if (!assoListProvider) throw new Error('asso list provider');
-                const pListId = assoListProvider.pApiId;
-                if (!pListId) throw new Error('pListId');
-
-                logger.log(`pListId ${pListId} processing assoLSPtoDelete`);
-
-                // first, disabling old ones
-                return assoLSPtoDelete.reduce((p, alsp, i) => {
-                  return p.then(() => {
-                    logger.log(`assoLSPtoDelete: ensureSyncCanProceed`);
-                    // first , we check if we can proceed
-                    return assoListProvider.ensureSyncCanProceed(syncId)
-                      .then(() => {
-                        logger.log(`assoLSPtoDelete: setStatusPendingActive`);
-                        return alsp.setStatusPendingActive();
-                      })
-                      .then(() => {
-                        logger.log(`assoLSPtoDelete: deleteSubscriber subscriberId=${alsp.subscriberId} email=${alsp.subscriber.referenceEmail}`);
-                        const subscriber = Mailer.Subscriber.loadFromModel(alsp.subscriber);
-                        return pApi.deleteSubscriber(pListId, subscriber.toISubscriber(alsp));
-                      })
-                      .then(() => {
-                        logger.log(`assoLSPtoDelete: setStatusActive`);
-                        return alsp.setStatusActive();
-                      })
-                      .then(() => {
-                        // notifying progress.
-                        if (i % 10 === 0) {
-                          const progress = 0.5 * (i + 1) / assoLSPtoDelete.length;
-                          logger.log(`assoLSPtoDelete: updatePApiStatus ${progress}`);
-                          return assoListProvider.updatePApiStatus({sync: { progress: progress }});
-                        }
-                      });
-                  });
-                }, Q())
-                .then(() => assoListProvider);
-              })
-              .then(assoListProvider => {
-                const pListId = assoListProvider.pApiId;
-
-                logger.log(`pListId ${pListId} processing assoLSPtoCreate`);
-
-                // second, activating new ones.
-                return assoLSPtoCreate.reduce((p, alsp, i) => {
-                  return p.then(() => {
-                    logger.log(`assoLSPtoCreate: ensureSyncCanProceed`);
-                    // first , we check if we can proceed
-                    return assoListProvider.ensureSyncCanProceed(syncId)
-                      .then(() => {
-                        logger.log(`assoLSPtoCreate: setStatusPendingActive`);
-                        return alsp.setStatusPendingUnsubscribed();
-                      })
-                      .then(() => {
-                        logger.log(`assoLSPtoCreate: createSubscriber subscriberId=${alsp.subscriberId} email=${alsp.subscriber.referenceEmail}`);
-                        const subscriber = Mailer.Subscriber.loadFromModel(alsp.subscriber);
-                        return pApi.createSubscriber(pListId, subscriber.toISubscriber(alsp))
-                          .then(iSubscriber => {
-                            if (!iSubscriber.get('id')) {
-                              throw new Error('no provider subscriber id!?');
-                            }
-                            return alsp.update({pApiId: iSubscriber.get('id')});
-                          });
-                      })
-                      .then(() => {
-                        logger.log(`assoLSPtoCreate: setStatusActive`);
-                        return alsp.setStatusActive();
-                      })
-                      .then(() => {
-                        // notifying progress.
-                        if (i % 10 === 0) {
-                          const progress = 0.5 * (i + 1) / assoLSPtoDelete.length;
-                          logger.log(`assoLSPtoCreate: updatePApiStatus ${progress}`);
-                          return assoListProvider.updatePApiStatus({sync: { progress: progress }});
-                        }
-                      });
-                  });
-                }, Q());
-              });
-          })
-          .then(
-            () => {
-              logger.log(`SUCCESS !`);
-              return this.stopSync(mailerProvider);
-            }
-          )
-          .catch(
-            err => {
-              /*
-               * catch-all worker status error
-               */
-              console.error(err.stack);
-              logger.log(`ERROR: ${err.message} ${JSON.stringify(err)}`);
-              assoListProvider.updatePApiStatus({ sync: { error: err.message }})
-                .then(() => this.stopSync(mailerProvider));
-            }
-          );
-
-          /*
-           *
-           *
-           *   W O R K E R       END
-           *
-           *
-           */
-        });
-      })
-      .then(() => {
-        return this.getSyncStatus(mailerProvider);
-      });
+    return sync.start(this, mailerProvider, {logger: logger});
   }
 
   stopSync(mailerProvider) {
@@ -638,14 +404,8 @@ class MailerList {
         );
       });
     }
-    const providerName = mailerProvider.getName();
-    logger.log(`[stopSync] provider ${providerName}`);
 
-    return this.getAssoProvider(mailerProvider)
-      .then(assoListProvider => {
-        logger.log(`[stopSync] assoListProvider found, updating PApiStatus to sync.id = null`);
-        return assoListProvider.updatePApiStatus({ sync: { id: null } });
-      });
+    return sync.stop(this, mailerProvider, {logger: logger});
   }
 
   getSyncStatus(mailerProvider) {
@@ -663,7 +423,7 @@ class MailerList {
     const providerName = mailerProvider.getName();
     logger.log(`[getSyncStatus] provider ${providerName}`);
 
-    return this.getAssoProvider(mailerProvider);
+    return sync.status(this, mailerProvider, {logger: logger});
   }
 
   toJSON() {
